Add server side logic for user dropdown
Refactored asset.service.js in client.
Before, a single method was used for both regular checking out and
checking out on behalf of someone. This was split into two different
methods; checkoutAssetForUser, and checkoutAsset.
In the details.controller.js the user information that is given to the
checkoutAssetForUser was changed to include the user's name in addition
to their email.
Refactored the checkForNewUser function to be in the usersService.js instead
of the authService.js.
Updated swagger documentation so that the client can send more
user info to the server instead of just the email when checking out a device
for someone else.

resolves #16