Add notification service
Added the notificationService.js with stubbed out methods for hipchat
and email. A user will specify which form of notifying they want in the
server config.
Added a helper method in the assetService for interfacing with the
notification service.
Added calls to the helper method throughout the assetService for
relaying a message to the notification service.

Resolves #31