Update to use current logged in user for Directory API
Retain legacy methods, but add new methods to allow for authentication using the current users access token.  Updated the existing server getGoogleUsers to allow for an optional token passed through from the client, but also added a new method on the client to bypass the server call and go to Google directly.  Work still needs to be done to integrate the Cordova app with these changes.